+++ jumi-launcher/src/main/java/fi/jumi/launcher/ui/PlainTextPrinter.java
@@ -4,14 +4,17 @@
 
 package fi.jumi.launcher.ui;
 
-import javax.annotation.concurrent.NotThreadSafe;
+import javax.annotation.concurrent.*;
 import java.io.IOException;
+
+import static fi.jumi.launcher.ui.PlainTextPrinter.Mode.*;
 
 @NotThreadSafe
 public class PlainTextPrinter implements Printer {
 
     private final Appendable out;
-    private boolean beginningOfLineOrInsideMeta = true;
+    private boolean beginningOfLine = true;
+    private Mode lastPrinted = Mode.STDOUT;
 
     public PlainTextPrinter(Appendable out) {
         this.out = out;
@@ -19,39 +22,46 @@ public class PlainTextPrinter implements Printer {
 
     @Override
     public void printOut(String text) {
+        beginOnNewLineUnlessAlreadyDoing(STDOUT);
         printTo(out, text);
-        printedNonMeta(text);
     }
 
     @Override
     public void printErr(String text) {
-        printTo(out, text);
-        printedNonMeta(text);
+        printOut(text);
     }
 
     @Override
-    public void printMeta(String text) {
-        if (!beginningOfLineOrInsideMeta) {
-            printOut("\n");
+    public void printMetaIncrement(String text) {
+        beginOnNewLineUnlessAlreadyDoing(META_INCREMENT);
+        printTo(out, text);
+    }
+
+    @Override
+    public void printMetaLine(String line) {
+        beginOnNewLineUnlessAlreadyDoing(META_LINE);
+        printTo(out, line);
+        printTo(out, "\n");
+    }
+
+    private void beginOnNewLineUnlessAlreadyDoing(Mode mode) {
+        if (!beginningOfLine && lastPrinted != mode) {
+            printTo(out, "\n");
         }
-        printTo(out, text);
-    }
-
-    @Override
-    public void printlnMeta(String line) {
-        printMeta(line);
-        printMeta("\n");
-    }
-
-    private void printedNonMeta(String text) {
-        beginningOfLineOrInsideMeta = text.endsWith("\n"); // matches both "\r\n" and "\n"
+        lastPrinted = mode;
     }
 
     private void printTo(Appendable target, String text) {
+        beginningOfLine = text.endsWith("\n"); // matches both "\r\n" and "\n"
         try {
             target.append(text);
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    @Immutable
+    enum Mode {
+        STDOUT, META_INCREMENT, META_LINE
     }
 }